feat: add recaw polling and chain switching for feed actions
- Add recaw pending state polling in Feed component to track recaw confirmations
- Add onRecawStateChange callback prop to FeedItem for parent notification
- Add SwitchChainModal check before like/recaw actions when on wrong chain
- Filter duplicate pending posts from feed when local pending post matches DB pending
- Add chain switching prompts to ReplyItem component for consistency

+++ solidity/contracts/CawActionsReplicator.sol
@@ -17,7 +17,7 @@ interface ICawActionsForReplicator {
     uint32 clientId;
     uint32 cawonce;
     uint32[] recipients;
-    uint128[] amounts;
+    uint64[] amounts;  // Whole CAW tokens (not wei)
     string text;
   }
